More thorough post content checking
Added checks for the `[post_content_filtered]` shortcode in the `post_content_filtered` field in order to add compatibility for certain Wordpress Themes, such as Layers. This addition is necessary in order for the affiliates page to work with Layers pages because layers pages have nothing in the `post_content` field and instead have everything in the `post_content_filtered` field.